sanei_thread: Draft fixup for sanei_thread issues.

+++ sanei/sanei_thread.c
@@ -117,16 +117,7 @@ sanei_thread_is_forked( void )
 static void
 sanei_thread_set_invalid( SANE_Pid *pid )
 {
-
-#ifdef WIN32
-#ifdef WINPTHREAD_API
-	*pid = (pthread_t) 0;
-#else
-	pid->p = 0;
-#endif
-#else
-	*pid = (pthread_t) -1;
-#endif
+  pid->pid = SANE_FALSE;
 }
 #endif
 
@@ -134,40 +125,34 @@ sanei_thread_set_invalid( SANE_Pid *pid )
 SANE_Bool
 sanei_thread_is_valid( SANE_Pid pid )
 {
-	SANE_Bool rc = SANE_TRUE;
-
-#ifdef WIN32
-#ifdef WINPTHREAD_API
-	if (pid == 0)
-#else
-	if (pid.p == 0)
-#endif
-	    rc = SANE_FALSE;
-#else
-	if (pid == (SANE_Pid) -1)
-	    rc = SANE_FALSE;
-#endif
-
-	return rc;
+    return pid.is_valid;
 }
 
 /* pthread_t is not an integer on all platform.  Do our best to return
  * a PID-like value from structure.  On platforms were it is an integer,
  * return that.
+ *
+ * Note: perhaps on platforms where pthread_t is a structure, compute a hash
+ * of the structure. It's not ideal but it is a thought.
+ *
  */
-static long
+long
 sanei_thread_pid_to_long( SANE_Pid pid )
 {
 	int rc;
 
+	if (!pid.is_valid)
+	{
+		return 0l;
+	}
 #ifdef WIN32
 #ifdef WINPTHREAD_API
-	rc = (long) pid;
+	rc = (long) pid.pid;
 #else
-	rc = pid.p;
+	rc = pid.pid.p;
 #endif
 #else
-	rc = (long) pid;
+	rc = (long) pid.pid;
 #endif
 
 	return rc;
@@ -176,18 +161,22 @@ sanei_thread_pid_to_long( SANE_Pid pid )
 int
 sanei_thread_kill( SANE_Pid pid )
 {
+	if (!pid.is_valid)
+	{
+		return -1;
+	}
 	DBG(2, "sanei_thread_kill() will kill %ld\n",
 	    sanei_thread_pid_to_long(pid));
 #ifdef USE_PTHREAD
 #if defined (__APPLE__) && defined (__MACH__)
-	return pthread_kill((pthread_t)pid, SIGUSR2);
+	return pthread_kill((pthread_t)pid.pid, SIGUSR2);
 #else
-	return pthread_cancel((pthread_t)pid);
+	return pthread_cancel((pthread_t)pid.pid);
 #endif
 #elif defined HAVE_OS2_H
-	return DosKillThread(pid);
+	return DosKillThread(pid.pid);
 #else
-	return kill( pid, SIGTERM );
+	return kill( pid.pid, SIGTERM );
 #endif
 }
 
@@ -208,39 +197,52 @@ local_thread( void *arg )
 /*
  * starts a new thread or process
  * parameters:
- * star  address of reader function
+ *
+ * func  address of reader function
  * args  pointer to scanner data structure
  *
  */
 SANE_Pid
 sanei_thread_begin( int (*func)(void *args), void* args )
 {
-	SANE_Pid pid;
+	SANE_Pid pid = {SANE_FALSE, -1};
 
 	td.func      = func;
 	td.func_data = args;
 
-	pid = _beginthread( local_thread, NULL, 1024*1024, (void*)&td );
-	if ( pid == -1 ) {
+	pid.pid = _beginthread( local_thread, NULL, 1024*1024, (void*)&td );
+	if ( pid.pid == -1 )
+	{
 		DBG( 1, "_beginthread() failed\n" );
-		return -1;
+		return pid;
 	}
 
 	DBG( 2, "_beginthread() created thread %d\n", pid );
+	pid.is_valid = SANE_TRUE;
+
 	return pid;
 }
 
 SANE_Pid
 sanei_thread_waitpid( SANE_Pid pid, int *status )
 {
-  if (status)
-    *status = 0;
-  return pid; /* DosWaitThread( (TID*) &pid, DCWW_WAIT);*/
+	if (!pid.is_valid)
+	{
+		return pid;
+	}
+
+	if (status)
+		*status = 0;
+	return pid; /* DosWaitThread( (TID*) &pid, DCWW_WAIT);*/
 }
 
 int
 sanei_thread_sendsig( SANE_Pid pid, int sig )
 {
+	if (!pid.is_valid)
+	{
+		return -1;
+	}
 	return 0;
 }
 
@@ -268,42 +270,60 @@ local_thread( void *arg )
 SANE_Pid
 sanei_thread_begin( int (*func)(void *args), void* args )
 {
-	SANE_Pid pid;
+	SANE_Pid pid = {SANE_FALSE, B_OK};
 
 	td.func      = func;
 	td.func_data = args;
 
-	pid = spawn_thread( local_thread, "sane thread (yes they can be)", B_NORMAL_PRIORITY, (void*)&td );
-	if ( pid < B_OK ) {
+	pid.pid = spawn_thread( local_thread, "sane thread (yes they can be)", B_NORMAL_PRIORITY, (void*)&td );
+	if ( pid.pid < B_OK ) {
 		DBG( 1, "spawn_thread() failed\n" );
-		return -1;
+		return pid;
 	}
-	if ( resume_thread(pid) < B_OK ) {
+	if ( resume_thread(pid.pid) < B_OK ) {
 		DBG( 1, "resume_thread() failed\n" );
-		return -1;
+		return pid;
 	}
 
 	DBG( 2, "spawn_thread() created thread %d\n", pid );
+	pid.is_valid = SANE_TRUE;
+
 	return pid;
 }
 
 SANE_Pid
 sanei_thread_waitpid( SANE_Pid pid, int *status )
 {
-  int32 st;
-  if ( wait_for_thread(pid, &st) < B_OK )
-    return -1;
-  if ( status )
-    *status = (int)st;
-  return pid;
+	int32 st;
+
+	if (!pid.is_valid)
+	{
+		return pid;
+	}
+
+	if ( wait_for_thread(pid.pid, &st) < B_OK )
+	{
+		pid.is_valid = SANE_FALSE;
+		return pid;
+	}
+	if ( status )
+		*status = (int)st;
+
+	return pid;
 }
 
 int
 sanei_thread_sendsig( SANE_Pid pid, int sig )
 {
+	if (!pid.is_valid)
+	{
+		return -1;
+	}
+
 	if (sig == SIGKILL)
 		sig = SIGKILLTHR;
-	return kill(pid, sig);
+
+	return kill(pid.pid, sig);
 }
 
 #else /* HAVE_OS2_H, __BEOS__ */
@@ -389,7 +409,7 @@ eval_wp_result( SANE_Pid pid, int wpres, int pf )
 {
 	int retval = SANE_STATUS_IO_ERROR;
 
-	if( wpres == pid ) {
+	if( wpres == pid.pid ) {
 
 		if( WIFEXITED(pf)) {
 			retval = WEXITSTATUS(pf);
@@ -413,7 +433,8 @@ sanei_thread_begin( int (func)(void *args), void* args )
 {
 #ifdef USE_PTHREAD
 	int result;
-	pthread_t thread;
+	SANE_Pid thread;
+
 #ifdef SIGPIPE
 	struct sigaction act;
 
@@ -434,27 +455,38 @@ sanei_thread_begin( int (func)(void *args), void* args )
 	td.func      = func;
 	td.func_data = args;
 
-	result = pthread_create( &thread, NULL, local_thread, &td );
+	result = pthread_create( &thread.pid, NULL, local_thread, &td );
 	usleep( 1 );
 
-	if ( result != 0 ) {
+	if ( result != 0 )
+	{
 		DBG( 1, "pthread_create() failed with %d\n", result );
 		sanei_thread_set_invalid(&thread);
 	}
 	else
+	{
 		DBG( 2, "pthread_create() created thread %ld\n",
 		     sanei_thread_pid_to_long(thread) );
-
-	return (SANE_Pid)thread;
-#else
-	SANE_Pid pid;
-	pid = fork();
-	if( pid < 0 ) {
-		DBG( 1, "fork() failed\n" );
-		return -1;
+		thread.is_valid = SANE_TRUE;
 	}
 
-	if( pid == 0 ) {
+	return thread;
+#else
+	SANE_Pid pid = {SANE_FALSE, 0};
+	pid.pid = fork();
+	if( pid.pid < 0 )
+	{
+		DBG( 1, "fork() failed\n" );
+		return pid;
+	}
+
+	pid.is_valid = SANE_TRUE;
+
+	/*
+	 * If I am the child....
+	 *
+	 */
+	if( pid.pid == 0 ) {
 
     	/* run in child context... */
 		int status = func( args );
@@ -473,10 +505,15 @@ sanei_thread_sendsig( SANE_Pid pid, int sig )
 {
 	DBG(2, "sanei_thread_sendsig() %d to thread (id=%ld)\n", sig,
 	    sanei_thread_pid_to_long(pid));
+
+	if (!pid.is_valid)
+	{
+		return -1;
+	}
 #ifdef USE_PTHREAD
-	return pthread_kill( (pthread_t)pid, sig );
+	return pthread_kill( (pthread_t)pid.pid, sig );
 #else
-	return kill( pid, sig );
+	return kill( pid.pid, sig );
 #endif
 }
 
@@ -491,32 +528,55 @@ sanei_thread_waitpid( SANE_Pid pid, int *status )
 	SANE_Pid result = pid;
 	int stat;
 
+	/*
+	 * Can't join if the PID is invalid.
+	 *
+	 * Normally, we would assume that the below call would fail if
+	 * the provided pid was invalid. However, we are now using a separate
+	 * boolean flag in SANE_Pid so we must check this.
+	 *
+	 * We must assume that the caller is making rational assumptions when
+	 * using SANE_Pid. You cannot assume that SANE_Pid is a pthread_t in
+	 * particular.
+	 *
+	 */
+	if (!pid.is_valid)
+	{
+		DBG(1, "sanei_thread_waitpid() - provided pid is invalid!\n");
+		return pid;
+	}
+
 	stat = 0;
 
-	DBG(2, "sanei_thread_waitpid() - %ld\n",
-	    sanei_thread_pid_to_long(pid));
+	DBG(2, "sanei_thread_waitpid() - %ld\n", sanei_thread_pid_to_long(pid));
 #ifdef USE_PTHREAD
 	int rc;
-	rc = pthread_join( (pthread_t)pid, (void*)&ls );
+	rc = pthread_join( pid.pid, (void*)&ls );
 
-	if( 0 == rc ) {
-		if( PTHREAD_CANCELED == ls ) {
+	if( 0 == rc )
+	{
+		if( PTHREAD_CANCELED == ls )
+		{
 			DBG(2, "* thread has been canceled!\n" );
 			stat = SANE_STATUS_GOOD;
-		} else {
+		}
+		else
+		{
 			stat = *ls;
 		}
 		DBG(2, "* result = %d (%p)\n", stat, (void*)status );
 		result = pid;
 	}
-	if ( EDEADLK == rc ) {
-		if ( (pthread_t)pid != pthread_self() ) {
+	if ( EDEADLK == rc )
+	{
+		if (!pthread_equal(pid.pid, pthread_self()))
+		{
 			/* call detach in any case to make sure that the thread resources
 			 * will be freed, when the thread has terminated
 			 */
 			DBG(2, "* detaching thread(%ld)\n",
 			    sanei_thread_pid_to_long(pid) );
-			pthread_detach((pthread_t)pid);
+			pthread_detach(pid.pid);
 		}
 	}
 	if (status)
@@ -524,7 +584,7 @@ sanei_thread_waitpid( SANE_Pid pid, int *status )
 
 	restore_sigpipe();
 #else
-	result = waitpid( pid, &ls, 0 );
+	result = waitpid( pid.pid, &ls, 0 );
 	if((result < 0) && (errno == ECHILD)) {
 		stat   = SANE_STATUS_GOOD;
 		result = pid;
@@ -551,14 +611,36 @@ sanei_thread_get_status( SANE_Pid pid )
 	int ls, stat, result;
 
 	stat = SANE_STATUS_IO_ERROR;
-	if( pid > 0 ) {
-
-		result = waitpid( pid, &ls, WNOHANG );
+	if (pid.is_valid && (pid.pid > 0) )
+	{
+		result = waitpid( pid.pid, &ls, WNOHANG );
 
 		stat = eval_wp_result( pid, result, ls );
 	}
 	return stat;
 #endif
 }
+
+/*
+ * Note: for the case of where the underlying system pid is neither
+ * a pointer nor an integer, we should use an appropriate platform function
+ * for comparing pids. We will have to take each case as it comes.
+ *
+ */
+SANE_Bool
+sanei_thread_pid_compare (SANE_Pid pid1, SANE_Pid pid2)
+{
+	if (!pid1.is_valid || !pid2.is_valid)
+	{
+		return SANE_FALSE;
+	}
+
+#if defined USE_PTHREAD
+	return pthread_equal(pid1.pid, pid2.pid)? SANE_TRUE: SANE_FALSE;
+#else
+	return pid1.pid == pid2.pid;
+#endif
+}
+
 
 /* END sanei_thread.c .......................................................*/
